brains/requirements.txt
Django==1.3
South==0.7.3
gevent==0.13.3
gunicorn==0.12.1
lxml==2.3
requests
django-celery
redis
